client/pyproject.toml
[tool.poetry]
name = "client"
version = "0.1.0"
description = ""
authors = ["Arjun Raghav A <47040002+arjunraghav@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
grpcio = "^1.54.2"
protobuf = "^4.23.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
